Add new schemas
