+++ app/hooks/useCreateEntity/useCreateTrack.ts
@@ -5,7 +5,12 @@ import md5 from 'md5';
 
 /* Internal dependencies */
 import { useAccount } from '~/hooks/useAccount/useAccount';
-import { MODULES, COMMANDS, FEEDBACK_MESSAGES } from './constants';
+import {
+  MODULES,
+  COMMANDS,
+  FEEDBACK_MESSAGES,
+  AUDIO_CREATE_SCHEMA,
+} from '../../constants/blockchain';
 import { waitFor } from '~/helpers/helpers';
 import { CHAIN_ID } from '~/constants/app';
 import {
@@ -15,7 +20,6 @@ import {
   PostTxResponse,
   Method,
 } from '~/context/socketContext/types';
-import { AUDIO_CREATE_SCHEMA } from './schemas';
 import { useWS } from '../useWS/useWS';
 import { ValidationStatus } from './types';
 import { validate } from './validator';
